test for redrizzled mosaics

+++ grizli/pipeline/auto_script.py
@@ -404,22 +404,29 @@ def go(root='j010311+131615', HOME_PATH='$PWD',
         msg = '\n\n### Redo visit-level mosaics and catalogs for fine alignment\n\n'
         utils.log_comment(utils.LOGFILE, msg, show_date=True, verbose=True)
         
+        keep_visits = []
+        
         for visit in visits:
-            visit_files = glob.glob(visit['product']+'_dr*')
-            visit_files += glob.glob(visit['product']+'*.cat.*')
+            
+            visit_files = glob.glob(visit['product']+'*.cat.*')
+            visit_files += glob.glob(visit['product']+'_dr*')
             visit_files += glob.glob(visit['product']+'*seg.fits*')
-            for file in visit_files:
-                os.remove(file)
+            
+            if len(visit_files) > 0:
+                keep_visits.append(visit)
+                for file in visit_files:
+                    os.remove(file)
+                
                 
         # Redrizzle visit-level mosaics and remake catalogs
-        prep.drizzle_overlaps(visits, check_overlaps=False, skysub=False,
+        prep.drizzle_overlaps(keep_visits, check_overlaps=False, skysub=False,
                               static=False, pixfrac=0.5, scale=None, 
                               final_wcs=False, fetch_flats=False,
                               final_rot=None,
                               include_saturated=True)
         
         # Make new catalogs
-        for visit in visits:
+        for visit in keep_visits:
             if len(visit['files']) == 0:
                 continue
